pyproject.toml
[project]
name = "bot-instante-infinito"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "pytelegrambotapi>=4.27.0",
    "python-dotenv>=1.1.0",
]
